venv-python3-10_requirements.txt
absl-py==2.3.1
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
autopep8==2.3.2
beautifulsoup4==4.9.3
beautifultable==1.1.0
bidict==0.23.1
black==25.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==7.1.2
comm==0.2.3
contourpy==1.3.2
CounterFit==0.1.4.dev2
counterfit-connection==0.1.0.dev5
counterfit-shims-grove==0.1.4.dev5
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dnspython==2.2.1
eventlet==0.33.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
Flask==1.1.4
Flask-SocketIO==5.2.0
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==1.1.0
jedi==0.19.2
Jinja2==2.11.3
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
lxml==4.6.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.0.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytokens==0.1.10
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.15.3
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.10.0.dev20250924
torchaudio==2.8.0.dev20250924
torchvision==0.25.0.dev20250924
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==1.0.1
wrapt==1.17.3
wsproto==1.2.0
yapf==0.43.0
